Control.Monad.Logic

Description

A backtracking, logic programming monad.

Adapted from the paper /Backtracking, Interleaving, and Terminating Monad Transformers/, by Oleg Kiselyov, Chung-chieh Shan, Daniel P. Friedman, Amr Sabry (http://okmij.org/ftp/papers/LogicT.pdf).

Documentation

module Control.Monad.Logic.Class

The Logic monad

type Logic = LogicT Identity

The basic Logic monad, for performing backtracking computations returning values of type a.

logic :: (forall r. (a -> r -> r) -> r -> r) -> Logic a

A smart constructor for Logic computations.

runLogic :: Logic a -> (a -> r -> r) -> r -> r

Runs a Logic computation with the specified initial success and failure continuations.

observe :: Logic a -> a

Extracts the first result from a Logic computation.

observeMany :: Int -> Logic a -> [a]

Extracts up to a given number of results from a Logic computation.

observeAll :: Logic a -> [a]

Extracts all results from a Logic computation.

The LogicT monad transformer

newtype LogicT m a

A monad transformer for performing backtracking computations layered over another monad m.

Constructors

LogicT

Fields

unLogicT :: forall r. (a -> m r -> m r) -> m r -> m r

runLogicT :: LogicT m a -> (a -> m r -> m r) -> m r -> m r

Runs a LogicT computation with the specified initial success and failure continuations.

observeT :: Monad m => LogicT m a -> m a

Extracts the first result from a LogicT computation, failing otherwise.

observeManyT :: Monad m => Int -> LogicT m a -> m [a]

Extracts up to a given number of results from a LogicT computation.

observeAllT :: Monad m => LogicT m a -> m [a]

Extracts all results from a LogicT computation.

module Control.Monad

module Control.Monad.Trans